allow negative latency

+++ client/controller.cpp
@@ -146,7 +146,7 @@ bool Controller::sendTimeSyncMessage(long after)
 }
 
 
-void Controller::start(const PcmDevice& pcmDevice, const std::string& host, size_t port, size_t latency)
+void Controller::start(const PcmDevice& pcmDevice, const std::string& host, size_t port, int latency)
 {
 	pcmDevice_ = pcmDevice;
 	latency_ = latency;
